+++ tests/test_router.py
@@ -2,7 +2,7 @@ import unittest
 
 import pytest
 
-from rivr.http import Request, Response
+from rivr import Request, Response
 from rivr.router import Resolver404, Router
 from rivr.views import View