+++ tests/test_views.py
@@ -1,6 +1,6 @@
 import unittest
 
-from rivr.http import Request, Response
+from rivr import Request, Response
 from rivr.views import RedirectView, View